refactor(sprints): move mock-data from sprints to global mocks directory The mock-data used in sprints.controller.spec.ts and sprints.service.spec.ts was moved to the global mocks directory. This change was made to improve the organization of the codebase, making it easier to locate and reuse mock data across different modules.
